LayoutTests/webgpu/whlsl/increment-setter.html

<!DOCTYPE html>
<html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>Test prefix/postfix.</title>
<script src="js/test-harness.js"></script>
<script src="../js/webgpu-functions.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
const whlslTests = {};

whlslTests.incrementSetter = async () =>
{
    let program = `
        struct Foo {
            int z;
        }
        int operator.w(Foo foo) {
            return 8;
        }
        Foo operator.w=(Foo foo, int value) {
            foo.z = value;
            return foo;
        }
        int bar() {
            Foo foo;
            foo.z = 3;
            int value = foo.w++;
            return value*100 + foo.z*10 + foo.w;
        }
    `;
    assert_equals(await callIntFunction(program, "bar", []), 898);
}

runTests(whlslTests);
</script>
</html>
